Moved process_inbound_command tests into a separate file
