+++ backend/tests/wallet_tests/services/test_fund_pull_pre_approval.py
@@ -359,891 +359,6 @@ def test_create_and_approve_while_command_already_exist_in_db():
         )
 
 
-def test_process_inbound_command_as_payer_with_incoming_pending_and_no_record_in_db(
-    mock_method,
-):
-    user = generate_mock_user()
-    address = generate_my_address(user)
-    biller_address = generate_address()
-
-    cmd = generate_funds_pull_pre_approval_command(
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.pending,
-    )
-    mock_method(
-        context.get().offchain_client,
-        "process_inbound_request",
-        will_return=cmd,
-    )
-
-    unused = b"Unused because process_inbound_request is mocked"
-    code, resp = process_inbound_command(address, unused)
-    assert code == 200
-    assert resp
-    command_in_db = get_command_by_id(FUNDS_PULL_PRE_APPROVAL_ID)
-    assert command_in_db.address == address
-    assert command_in_db.biller_address == biller_address
-
-
-def test_process_inbound_command_as_payer_with_incoming_pending_while_record_db_exist(
-    mock_method,
-):
-    user = generate_mock_user()
-    address = generate_my_address(user)
-    biller_address = generate_address()
-
-    OneFundsPullPreApproval.run(
-        db_session=db_session,
-        account_id=user.account_id,
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.pending,
-        max_cumulative_unit="week",
-        max_cumulative_unit_value=1,
-        role=Role.PAYER,
-    )
-
-    cmd = generate_funds_pull_pre_approval_command(
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        max_cumulative_unit="month",
-        max_cumulative_unit_value=2,
-    )
-    mock_method(
-        context.get().offchain_client,
-        "process_inbound_request",
-        will_return=cmd,
-    )
-
-    unused = b"Unused because process_inbound_request is mocked"
-    code, resp = process_inbound_command(address, unused)
-    assert code == 200
-    assert resp
-    command_in_db = get_command_by_id(FUNDS_PULL_PRE_APPROVAL_ID)
-    assert command_in_db.address == address
-    assert command_in_db.biller_address == biller_address
-    assert command_in_db.max_cumulative_unit == "month"
-    assert command_in_db.max_cumulative_unit_value == 2
-
-
-def test_process_inbound_command_as_payer_with_incoming_reject_and_no_record_in_db(
-    mock_method,
-):
-    user = generate_mock_user()
-    address = generate_my_address(user)
-    biller_address = generate_address()
-
-    cmd = generate_funds_pull_pre_approval_command(
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.rejected,
-    )
-    mock_method(
-        context.get().offchain_client,
-        "process_inbound_request",
-        will_return=cmd,
-    )
-
-    with pytest.raises(FundsPullPreApprovalStateError):
-        unused = b"Unused because process_inbound_request is mocked"
-        process_inbound_command(address, unused)
-
-
-def test_process_inbound_command_as_payer_with_incoming_reject_while_record_db_exist(
-    mock_method,
-):
-    user = generate_mock_user()
-    address = generate_my_address(user)
-    biller_address = generate_address()
-
-    OneFundsPullPreApproval.run(
-        db_session=db_session,
-        account_id=user.account_id,
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.pending,
-        role=Role.PAYER,
-    )
-
-    cmd = generate_funds_pull_pre_approval_command(
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.rejected,
-    )
-    mock_method(
-        context.get().offchain_client,
-        "process_inbound_request",
-        will_return=cmd,
-    )
-
-    with pytest.raises(FundsPullPreApprovalStateError):
-        unused = b"Unused because process_inbound_request is mocked"
-        process_inbound_command(address, unused)
-
-
-def test_process_inbound_command_as_payer_with_incoming_valid_and_no_record_in_db(
-    mock_method,
-):
-    user = generate_mock_user()
-    address = generate_my_address(user)
-    biller_address = generate_address()
-
-    cmd = generate_funds_pull_pre_approval_command(
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.valid,
-    )
-    mock_method(
-        context.get().offchain_client,
-        "process_inbound_request",
-        will_return=cmd,
-    )
-
-    with pytest.raises(FundsPullPreApprovalStateError):
-        unused = b"Unused because process_inbound_request is mocked"
-        process_inbound_command(address, unused)
-
-
-def test_process_inbound_command_as_payer_with_incoming_valid_while_record_db_exist(
-    mock_method,
-):
-    user = generate_mock_user()
-    address = generate_my_address(user)
-    biller_address = generate_address()
-
-    OneFundsPullPreApproval.run(
-        db_session=db_session,
-        account_id=user.account_id,
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.valid,
-        role=Role.PAYER,
-    )
-
-    cmd = generate_funds_pull_pre_approval_command(
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.rejected,
-    )
-    mock_method(
-        context.get().offchain_client,
-        "process_inbound_request",
-        will_return=cmd,
-    )
-
-    with pytest.raises(FundsPullPreApprovalStateError):
-        unused = b"Unused because process_inbound_request is mocked"
-        process_inbound_command(address, unused)
-
-
-def test_process_inbound_command_as_payer_with_incoming_closed_and_no_record_in_db(
-    mock_method,
-):
-    user = generate_mock_user()
-    address = generate_my_address(user)
-    biller_address = generate_address()
-
-    cmd = generate_funds_pull_pre_approval_command(
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.closed,
-    )
-    mock_method(
-        context.get().offchain_client,
-        "process_inbound_request",
-        will_return=cmd,
-    )
-
-    with pytest.raises(FundsPullPreApprovalStateError):
-        unused = b"Unused because process_inbound_request is mocked"
-        process_inbound_command(address, unused)
-
-
-def test_process_inbound_command_as_payer_with_incoming_closed_while_record_db_exist_with_valid_status(
-    mock_method,
-):
-    user = generate_mock_user()
-    address = generate_my_address(user)
-    biller_address = generate_address()
-
-    OneFundsPullPreApproval.run(
-        db_session=db_session,
-        account_id=user.account_id,
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.valid,
-        role=Role.PAYER,
-    )
-
-    cmd = generate_funds_pull_pre_approval_command(
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.closed,
-    )
-    mock_method(
-        context.get().offchain_client,
-        "process_inbound_request",
-        will_return=cmd,
-    )
-
-    unused = b"Unused because process_inbound_request is mocked"
-    code, resp = process_inbound_command(address, unused)
-    assert code == 200
-    assert resp
-    command_in_db = get_command_by_id(FUNDS_PULL_PRE_APPROVAL_ID)
-    assert command_in_db.address == address
-    assert command_in_db.biller_address == biller_address
-    assert command_in_db.status == FundPullPreApprovalStatus.closed
-
-
-def test_process_inbound_command_as_payer_with_incoming_closed_while_record_db_exist_with_pending_status(
-    mock_method,
-):
-    user = generate_mock_user()
-    address = generate_my_address(user)
-    biller_address = generate_address()
-
-    OneFundsPullPreApproval.run(
-        db_session=db_session,
-        account_id=user.account_id,
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.pending,
-        role=Role.PAYER,
-    )
-
-    cmd = generate_funds_pull_pre_approval_command(
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.closed,
-    )
-    mock_method(
-        context.get().offchain_client,
-        "process_inbound_request",
-        will_return=cmd,
-    )
-
-    unused = b"Unused because process_inbound_request is mocked"
-    code, resp = process_inbound_command(address, unused)
-    assert code == 200
-    assert resp
-    command_in_db = get_command_by_id(FUNDS_PULL_PRE_APPROVAL_ID)
-    assert command_in_db.address == address
-    assert command_in_db.biller_address == biller_address
-    assert command_in_db.status == FundPullPreApprovalStatus.closed
-
-
-def test_process_inbound_command_as_payer_with_incoming_closed_while_record_db_exist_with_rejected_status(
-    mock_method,
-):
-    user = generate_mock_user()
-    address = generate_my_address(user)
-    biller_address = generate_address()
-
-    OneFundsPullPreApproval.run(
-        db_session=db_session,
-        account_id=user.account_id,
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.rejected,
-        role=Role.PAYER,
-    )
-
-    cmd = generate_funds_pull_pre_approval_command(
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.closed,
-    )
-    mock_method(
-        context.get().offchain_client,
-        "process_inbound_request",
-        will_return=cmd,
-    )
-
-    with pytest.raises(FundsPullPreApprovalStateError):
-        unused = b"Unused because process_inbound_request is mocked"
-        process_inbound_command(address, unused)
-
-
-def test_process_inbound_command_as_payee_with_incoming_pending_and_no_record_in_db(
-    mock_method,
-):
-    address = generate_address()
-    biller_user = generate_mock_user()
-    biller_address = generate_my_address(biller_user)
-
-    cmd = generate_funds_pull_pre_approval_command(
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.pending,
-    )
-    mock_method(
-        context.get().offchain_client,
-        "process_inbound_request",
-        will_return=cmd,
-    )
-
-    with pytest.raises(FundsPullPreApprovalStateError):
-        unused = b"Unused because process_inbound_request is mocked"
-        process_inbound_command(address, unused)
-
-
-def test_process_inbound_command_as_payee_with_incoming_pending_while_record_db_exist(
-    mock_method,
-):
-    address = generate_address()
-    biller_user = generate_mock_user()
-    biller_address = generate_my_address(biller_user)
-
-    OneFundsPullPreApproval.run(
-        db_session=db_session,
-        account_id=biller_user.account_id,
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.pending,
-        max_cumulative_unit="week",
-        max_cumulative_unit_value=1,
-        role=Role.PAYEE,
-    )
-
-    cmd = generate_funds_pull_pre_approval_command(
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.pending,
-        max_cumulative_unit="month",
-        max_cumulative_unit_value=2,
-    )
-    mock_method(
-        context.get().offchain_client,
-        "process_inbound_request",
-        will_return=cmd,
-    )
-
-    with pytest.raises(FundsPullPreApprovalStateError):
-        unused = b"Unused because process_inbound_request is mocked"
-        process_inbound_command(address, unused)
-
-
-def test_process_inbound_command_as_payee_with_incoming_reject_and_no_record_in_db(
-    mock_method,
-):
-    address = generate_address()
-    biller_user = generate_mock_user()
-    biller_address = generate_my_address(biller_user)
-
-    cmd = generate_funds_pull_pre_approval_command(
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.rejected,
-    )
-    mock_method(
-        context.get().offchain_client,
-        "process_inbound_request",
-        will_return=cmd,
-    )
-
-    with pytest.raises(FundsPullPreApprovalStateError):
-        unused = b"Unused because process_inbound_request is mocked"
-        process_inbound_command(address, unused)
-
-
-def test_process_inbound_command_as_payee_with_incoming_reject_while_record_db_exist_with_pending_status(
-    mock_method,
-):
-    address = generate_address()
-    biller_user = generate_mock_user()
-    biller_address = generate_my_address(biller_user)
-
-    OneFundsPullPreApproval.run(
-        db_session=db_session,
-        account_id=biller_user.account_id,
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.pending,
-        role=Role.PAYEE,
-    )
-
-    cmd = generate_funds_pull_pre_approval_command(
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.rejected,
-    )
-    mock_method(
-        context.get().offchain_client,
-        "process_inbound_request",
-        will_return=cmd,
-    )
-
-    unused = b"Unused because process_inbound_request is mocked"
-    code, resp = process_inbound_command(address, unused)
-    assert code == 200
-    assert resp
-    command_in_db = get_command_by_id(FUNDS_PULL_PRE_APPROVAL_ID)
-    assert command_in_db.address == address
-    assert command_in_db.biller_address == biller_address
-    assert command_in_db.status == FundPullPreApprovalStatus.rejected
-
-
-def test_process_inbound_command_as_payee_with_incoming_reject_while_record_db_exist_with_valid_status(
-    mock_method,
-):
-    address = generate_address()
-    biller_user = generate_mock_user()
-    biller_address = generate_my_address(biller_user)
-
-    OneFundsPullPreApproval.run(
-        db_session=db_session,
-        account_id=biller_user.account_id,
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.valid,
-        role=Role.PAYEE,
-    )
-
-    cmd = generate_funds_pull_pre_approval_command(
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.rejected,
-    )
-    mock_method(
-        context.get().offchain_client,
-        "process_inbound_request",
-        will_return=cmd,
-    )
-
-    with pytest.raises(FundsPullPreApprovalStateError):
-        unused = b"Unused because process_inbound_request is mocked"
-        process_inbound_command(address, unused)
-
-
-def test_process_inbound_command_as_payee_with_incoming_reject_while_record_db_exist_with_rejected_status(
-    mock_method,
-):
-    address = generate_address()
-    biller_user = generate_mock_user()
-    biller_address = generate_my_address(biller_user)
-
-    OneFundsPullPreApproval.run(
-        db_session=db_session,
-        account_id=biller_user.account_id,
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.rejected,
-        role=Role.PAYEE,
-    )
-
-    cmd = generate_funds_pull_pre_approval_command(
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.rejected,
-    )
-    mock_method(
-        context.get().offchain_client,
-        "process_inbound_request",
-        will_return=cmd,
-    )
-
-    with pytest.raises(FundsPullPreApprovalStateError):
-        unused = b"Unused because process_inbound_request is mocked"
-        process_inbound_command(address, unused)
-
-
-def test_process_inbound_command_as_payee_with_incoming_reject_while_record_db_exist_with_closed_status(
-    mock_method,
-):
-    address = generate_address()
-    biller_user = generate_mock_user()
-    biller_address = generate_my_address(biller_user)
-
-    OneFundsPullPreApproval.run(
-        db_session=db_session,
-        account_id=biller_user.account_id,
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.closed,
-        role=Role.PAYEE,
-    )
-
-    cmd = generate_funds_pull_pre_approval_command(
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.rejected,
-    )
-    mock_method(
-        context.get().offchain_client,
-        "process_inbound_request",
-        will_return=cmd,
-    )
-
-    with pytest.raises(FundsPullPreApprovalStateError):
-        unused = b"Unused because process_inbound_request is mocked"
-        process_inbound_command(address, unused)
-
-
-def test_process_inbound_command_as_payee_with_incoming_valid_and_no_record_in_db(
-    mock_method,
-):
-    address = generate_address()
-    biller_user = generate_mock_user()
-    biller_address = generate_my_address(biller_user)
-
-    cmd = generate_funds_pull_pre_approval_command(
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.valid,
-    )
-    mock_method(
-        context.get().offchain_client,
-        "process_inbound_request",
-        will_return=cmd,
-    )
-
-    with pytest.raises(FundsPullPreApprovalStateError):
-        unused = b"Unused because process_inbound_request is mocked"
-        process_inbound_command(address, unused)
-
-
-def test_process_inbound_command_as_payee_with_incoming_valid_while_record_db_exist_with_pending_status(
-    mock_method,
-):
-    address = generate_address()
-    biller_user = generate_mock_user()
-    biller_address = generate_my_address(biller_user)
-
-    OneFundsPullPreApproval.run(
-        db_session=db_session,
-        account_id=biller_user.account_id,
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.pending,
-        role=Role.PAYEE,
-    )
-
-    cmd = generate_funds_pull_pre_approval_command(
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.valid,
-    )
-    mock_method(
-        context.get().offchain_client,
-        "process_inbound_request",
-        will_return=cmd,
-    )
-
-    unused = b"Unused because process_inbound_request is mocked"
-    code, resp = process_inbound_command(address, unused)
-    assert code == 200
-    assert resp
-    command_in_db = get_command_by_id(FUNDS_PULL_PRE_APPROVAL_ID)
-    assert command_in_db.address == address
-    assert command_in_db.biller_address == biller_address
-    assert command_in_db.status == FundPullPreApprovalStatus.valid
-
-
-def test_process_inbound_command_as_payee_with_incoming_valid_while_record_db_exist_with_valid_status(
-    mock_method,
-):
-    address = generate_address()
-    biller_user = generate_mock_user()
-    biller_address = generate_my_address(biller_user)
-
-    OneFundsPullPreApproval.run(
-        db_session=db_session,
-        account_id=biller_user.account_id,
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.valid,
-        role=Role.PAYEE,
-    )
-
-    cmd = generate_funds_pull_pre_approval_command(
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.valid,
-    )
-    mock_method(
-        context.get().offchain_client,
-        "process_inbound_request",
-        will_return=cmd,
-    )
-
-    with pytest.raises(FundsPullPreApprovalStateError):
-        unused = b"Unused because process_inbound_request is mocked"
-        process_inbound_command(address, unused)
-
-
-def test_process_inbound_command_as_payee_with_incoming_valid_while_record_db_exist_with_rejected_status(
-    mock_method,
-):
-    address = generate_address()
-    biller_user = generate_mock_user()
-    biller_address = generate_my_address(biller_user)
-
-    OneFundsPullPreApproval.run(
-        db_session=db_session,
-        account_id=biller_user.account_id,
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.rejected,
-        role=Role.PAYEE,
-    )
-
-    cmd = generate_funds_pull_pre_approval_command(
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.valid,
-    )
-    mock_method(
-        context.get().offchain_client,
-        "process_inbound_request",
-        will_return=cmd,
-    )
-
-    with pytest.raises(FundsPullPreApprovalStateError):
-        unused = b"Unused because process_inbound_request is mocked"
-        process_inbound_command(address, unused)
-
-
-def test_process_inbound_command_as_payee_with_incoming_valid_while_record_db_exist_with_closed_status(
-    mock_method,
-):
-    address = generate_address()
-    biller_user = generate_mock_user()
-    biller_address = generate_my_address(biller_user)
-
-    OneFundsPullPreApproval.run(
-        db_session=db_session,
-        account_id=biller_user.account_id,
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.closed,
-        role=Role.PAYEE,
-    )
-
-    cmd = generate_funds_pull_pre_approval_command(
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.valid,
-    )
-    mock_method(
-        context.get().offchain_client,
-        "process_inbound_request",
-        will_return=cmd,
-    )
-
-    with pytest.raises(FundsPullPreApprovalStateError):
-        unused = b"Unused because process_inbound_request is mocked"
-        process_inbound_command(address, unused)
-
-
-def test_process_inbound_command_as_payee_with_incoming_closed_and_no_record_in_db(
-    mock_method,
-):
-    address = generate_address()
-    biller_user = generate_mock_user()
-    biller_address = generate_my_address(biller_user)
-
-    cmd = generate_funds_pull_pre_approval_command(
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.closed,
-    )
-    mock_method(
-        context.get().offchain_client,
-        "process_inbound_request",
-        will_return=cmd,
-    )
-
-    with pytest.raises(FundsPullPreApprovalStateError):
-        unused = b"Unused because process_inbound_request is mocked"
-        process_inbound_command(address, unused)
-
-
-def test_process_inbound_command_as_payee_with_incoming_closed_while_record_db_exist_with_valid_status(
-    mock_method,
-):
-    address = generate_address()
-    biller_user = generate_mock_user()
-    biller_address = generate_my_address(biller_user)
-
-    OneFundsPullPreApproval.run(
-        db_session=db_session,
-        account_id=biller_user.account_id,
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.valid,
-        role=Role.PAYEE,
-    )
-
-    cmd = generate_funds_pull_pre_approval_command(
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.closed,
-    )
-    mock_method(
-        context.get().offchain_client,
-        "process_inbound_request",
-        will_return=cmd,
-    )
-
-    unused = b"Unused because process_inbound_request is mocked"
-    code, resp = process_inbound_command(address, unused)
-    assert code == 200
-    assert resp
-    command_in_db = get_command_by_id(FUNDS_PULL_PRE_APPROVAL_ID)
-    assert command_in_db.address == address
-    assert command_in_db.biller_address == biller_address
-    assert command_in_db.status == FundPullPreApprovalStatus.closed
-
-
-def test_process_inbound_command_as_payee_with_incoming_closed_while_record_db_exist_with_pending_status(
-    mock_method,
-):
-    address = generate_address()
-    biller_user = generate_mock_user()
-    biller_address = generate_my_address(biller_user)
-
-    OneFundsPullPreApproval.run(
-        db_session=db_session,
-        account_id=biller_user.account_id,
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.pending,
-        role=Role.PAYEE,
-    )
-
-    cmd = generate_funds_pull_pre_approval_command(
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.closed,
-    )
-    mock_method(
-        context.get().offchain_client,
-        "process_inbound_request",
-        will_return=cmd,
-    )
-
-    unused = b"Unused because process_inbound_request is mocked"
-    code, resp = process_inbound_command(address, unused)
-    assert code == 200
-    assert resp
-    command_in_db = get_command_by_id(FUNDS_PULL_PRE_APPROVAL_ID)
-    assert command_in_db.address == address
-    assert command_in_db.biller_address == biller_address
-    assert command_in_db.status == FundPullPreApprovalStatus.closed
-
-
-def test_process_inbound_command_as_payee_with_incoming_closed_while_record_db_exist_with_rejected_status(
-    mock_method,
-):
-    address = generate_address()
-    biller_user = generate_mock_user()
-    biller_address = generate_my_address(biller_user)
-
-    OneFundsPullPreApproval.run(
-        db_session=db_session,
-        account_id=biller_user.account_id,
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.rejected,
-        role=Role.PAYEE,
-    )
-
-    cmd = generate_funds_pull_pre_approval_command(
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.closed,
-    )
-    mock_method(
-        context.get().offchain_client,
-        "process_inbound_request",
-        will_return=cmd,
-    )
-
-    with pytest.raises(FundsPullPreApprovalStateError):
-        unused = b"Unused because process_inbound_request is mocked"
-        process_inbound_command(address, unused)
-
-
-def test_process_inbound_command_as_payee_with_incoming_closed_while_record_db_exist_with_closed_status(
-    mock_method,
-):
-    address = generate_address()
-    biller_user = generate_mock_user()
-    biller_address = generate_my_address(biller_user)
-
-    OneFundsPullPreApproval.run(
-        db_session=db_session,
-        account_id=biller_user.account_id,
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.closed,
-        role=Role.PAYEE,
-    )
-
-    cmd = generate_funds_pull_pre_approval_command(
-        address=address,
-        biller_address=biller_address,
-        funds_pull_pre_approval_id=FUNDS_PULL_PRE_APPROVAL_ID,
-        status=FundPullPreApprovalStatus.closed,
-    )
-    mock_method(
-        context.get().offchain_client,
-        "process_inbound_request",
-        will_return=cmd,
-    )
-
-    with pytest.raises(FundsPullPreApprovalStateError):
-        unused = b"Unused because process_inbound_request is mocked"
-        process_inbound_command(address, unused)
-
-
 def test_process_inbound_command_as_both__happy_flow(mock_method):
     payer_user = generate_mock_user(user_name="payer_user")
     payer_bech32 = generate_my_address(payer_user)
